\name{hamilton}
          \docType{data}
          \alias{hamilton}
          \title{Example dose-response curves from Hamilton (1977)}
          \description{
	  These are the example dose-response curves given in Hamilton (1977). 
	  Note that, as per Hamilton (1978), the confidence intervals given in 
	  Hamilton (1977) for these data sets are incorrect.
         }
          \usage{hamilton}
          \format{A list containing ten data frames: 
	  	    dr1a, dr1b, dr1c, dr1d, dr1e, 
	  	    dr4a, dr4b, dr4c, dr4d, dr4e, }
          \source{Hamilton, 1977.}
	  \references{Hamilton, M. A.; Russo, R. C.; Thurston, R. V. Trimmed 
  Spearman-Karber Method for Estimating Median Lethal Concentrations 
  in Toxicity Bioassays. Enviro. Sci. Tech. 1977, 11 (7), 714-719. 
  \url{http://dx.doi.org/10.1021/es60130a004}

  Ibid, 1978, 12 (4), 417. \url{http://dx.doi.org/10.1021/es60140a017}}
          \keyword{datasets}
